+++ test/index_test.js
@@ -197,6 +197,77 @@ describe("Objectware", function() {
     })
   })
 
+  describe(".defaults", function() {
+    it("must return undefined given nothing", function() {
+      demand(_.defaults()).be.undefined()
+    })
+
+    it("must return null given null", function() {
+      demand(_.defaults(null)).be.null()
+    })
+
+    it("must return undefined given undefined and a source", function() {
+      demand(_.defaults(undefined, {name: "John"})).be.undefined()
+    })
+
+    it("must return null given null and a source", function() {
+      demand(_.defaults(null, {name: "John"})).be.null()
+    })
+
+    it("must return target given no source", function() {
+      var obj = {}
+      _.defaults(obj).must.equal(obj)
+    })
+
+    it("must return target given one source", function() {
+      var obj = {}
+      _.defaults(obj, {name: "John"}).must.equal(obj)
+    })
+
+    it("must assign properties to target from one source", function() {
+      _.defaults({}, {name: "John"}).must.eql({name: "John"})
+    })
+
+    it("must assign properties to target from two sources", function() {
+      _.defaults({}, {name: "John"}, {age: 13}).must.eql({name: "John", age: 13})
+    })
+
+    it("must not assign properties in Object.prototype", function() {
+      _.defaults({}, {hasOwnProperty: true}).must.eql({})
+    })
+
+    it("must assign properties in Object.prototype when inheriting from null",
+      function() {
+      var obj = _.defaults(Object.create(null), {hasOwnProperty: true})
+      demand(obj).eql({hasOwnProperty: true})
+    })
+
+    it("must not overwrite property if target already has it", function() {
+      _.defaults({name: "John"}, {name: "Mike"}).must.eql({name: "John"})
+    })
+
+    it("must not overwrite property with later source", function() {
+      _.defaults({}, {name: "John"}, {name: "Mike"}).must.eql({name: "John"})
+    })
+
+    it("must not change target given no source", function() {
+      _.defaults({name: "John"}).must.eql({name: "John"})
+    })
+
+    it("must assign properties from inherited sources", function() {
+      _.defaults({}, Object.create({name: "John"})).must.eql({name: "John"})
+    })
+
+    it("must not assign unenumerable properties", function() {
+      var source = Object.defineProperty({}, "name", {value: "John"})
+      _.defaults({}, source).must.eql({})
+    })
+
+    it("must assign properties with undefined value", function() {
+      _.defaults({}, {name: undefined}).must.eql({name: undefined})
+    })
+  })
+
   describe(".each", function() {
     it("must call function with value, key and object", function() {
       var obj = {name: "John", age: 42, height: 190}