+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Unreleased
 - Adds [`Objectware.create`][create].
+- Adds [`Objectware.defaults`][defaults].
 
 [create]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.create
+[defaults]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.defaults
 
 ## 1.5.0 (Feb 10, 2015)
 - Adds [`Objectware.each`][each].